requirements.txt
kafka==1.3.5
kafka-python==2.0.2
requests==2.27.1
json5==0.9.6
pyspark==3.4.1
cassandra-driver==3.29.0
python-telegram-bot==20.7
